% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmd.R
\name{bfmd.pdf}
\alias{bfmd.pdf}
\title{PDF of the Bulk FMD}
\usage{
bfmd.pdf(m, theta)
}
\arguments{
\item{m}{a numeric vector of earthquake magnitudes}

\item{theta}{a list of 3 parameters:
\itemize{
\item \code{beta}  the Gutenberg-Richter parameter value
\item \code{mu}    the mean value of the cumulative normal distribution
\item \code{sigma} the standard deviation of the cumulative normal distribution
}}
}
\value{
A numeric vector of densities.
}
\description{
Compute the probability density function (PDF) of the bulk
frenquency-magnitude distribution (FMD), as defined by Ogata and Katsura (2006)
(see also Ringdal, 1975; Ogata and Katsura, 1993).
}
\details{
The bulk FMD model is the product of the Gutenberg-Richter model and a detection function
defined as the cumulative Normal distribution. Its FMD has a curved shape in the log-lin
space and corresponds to the case where the completeness magnitude
\out{m<sub>c</sub>} is variable (the FMD curvature representing the
\out{m<sub>c</sub>} distribution; read more in Mignan, 2012; Mignan and Chen, 2016).
}
\references{
Mignan, A. (2012), Functional shape of the earthquake frequency-magnitude
distribution and completeness magnitude, J. Geophys. Res., 117, B08302,
\href{http://onlinelibrary.wiley.com/doi/10.1029/2012JB009347/full}{doi: 10.1029/2012JB009347}

Mignan, A., Chen, C.-C. (2016), The Spatial Scale of Detected Seismicity,
Pure Appl. Geophys., 173, 117-124,
\href{https://link.springer.com/article/10.1007/s00024-015-1133-7}{doi: 10.1007/s00024-015-1133-7}

Ogata, Y., Katsura, K. (1993), Analysis of temporal and spatial
heterogeneity of magnitude frequency distribution inferred from earthquake
catalogues, Geophys. J. Int.,
\href{http://onlinelibrary.wiley.com/doi/10.1111/j.1365-246X.1993.tb04663.x/abstract}{113, 727-738}

Ogata, Y., Katsura, K. (2006), Immediate and updated forecasting of
aftershock hazard, Geophys. Res. Lett., 33, L10305,
\href{http://onlinelibrary.wiley.com/doi/10.1029/2006GL025888/full}{doi: 10.1029/2006GL025888}

Ringdal, F. (1975), On the estimation of seismic detection thresholds,
Bull. Seismol. Soc. Am.,
\href{https://pubs.geoscienceworld.org/ssa/bssa/article-abstract/65/6/1631/117495/on-the-estimation-of-seismic-detection-thresholds}{65, 1631-1642}
}
\seealso{
\code{beta.mle}; \code{bfmd.sim}
}
