% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmd.R
\name{bfmd.sim}
\alias{bfmd.sim}
\title{Simulation of the Bulk FMD}
\usage{
bfmd.sim(N, theta, mbin = 0.1, mmin = 0, mmax = 9)
}
\arguments{
\item{N}{the approximate number of earthquakes to simulate}

\item{theta}{a list of 3 parameters:
\itemize{
\item \code{beta}  the Gutenberg-Richter parameter value
\item \code{mu}    the mean value of the cumulative normal distribution
\item \code{sigma} the standard deviation of the cumulative normal distribution
}}

\item{mbin}{the magnitude binning value (if not provided, \code{mbin = 0.1})}

\item{mmin}{the minimum magnitude value (if not provided, \code{mmin = 0})}

\item{mmax}{the maximum magnitude value (if not provided, \code{mmax = 9})}
}
\value{
A numeric vector of approximatively \code{N} earthquake magnitudes.
}
\description{
Simulate the bulk earthquake frequency magnitude distribution (bFMD) by
applying the Thinning Method (Lewis and Shedler, 1979) to the curved FMD model of Ringdal (1975);
Ogata and Katsura (1993; 2006).
}
\details{
The bulk FMD model is the product of the Gutenberg-Richter model and a detection function
defined as the cumulative Normal distribution. Its FMD has a curved shape in the log-lin
space and corresponds to the case where the completeness magnitude
\emph{\out{m<sub>c</sub>}} is variable (the FMD curvature representing the
\emph{\out{m<sub>c</sub>}} distribution; read more in Mignan, 2012; Mignan and Chen, 2016).
}
\examples{
theta <- list(beta = log(10), mu = 2, sigma = 0.3)
m.sim <- bfmd.sim(1e3, theta)
mdistr <- fmd(m.sim)
plot(mdistr$mi, mdistr$Ni, log = "y")
points(mdistr$mi, mdistr$ni)
}
\references{
Lewis, P.A.W., Shedler, G.S. (1979), Simulation of nonhomogeneous poisson
processes by thinning, Naval Res. Logistics, 26, 403-413
\href{http://onlinelibrary.wiley.com/doi/10.1002/nav.3800260304/abstract}{doi: 10.1002/nav.3800260304}

Mignan, A. (2012), Functional shape of the earthquake frequency-magnitude
distribution and completeness magnitude, J. Geophys. Res., 117, B08302,
\href{http://onlinelibrary.wiley.com/doi/10.1029/2012JB009347/full}{doi: 10.1029/2012JB009347}

Mignan, A., Chen, C.-C. (2016), The Spatial Scale of Detected Seismicity,
Pure Appl. Geophys., 173, 117-124,
\href{https://link.springer.com/article/10.1007/s00024-015-1133-7}{doi: 10.1007/s00024-015-1133-7}

Ogata, Y., Katsura, K. (1993), Analysis of temporal and spatial
heterogeneity of magnitude frequency distribution inferred from earthquake
catalogues, Geophys. J. Int.,
\href{http://onlinelibrary.wiley.com/doi/10.1111/j.1365-246X.1993.tb04663.x/abstract}{113, 727-738}

Ogata, Y., Katsura, K. (2006), Immediate and updated forecasting of
aftershock hazard, Geophys. Res. Lett., 33, L10305,
\href{http://onlinelibrary.wiley.com/doi/10.1029/2006GL025888/full}{doi: 10.1029/2006GL025888}

Ringdal, F. (1975), On the estimation of seismic detection thresholds,
Bull. Seismol. Soc. Am.,
\href{https://pubs.geoscienceworld.org/ssa/bssa/article-abstract/65/6/1631/117495/on-the-estimation-of-seismic-detection-thresholds}{65, 1631-1642}
}
\seealso{
\code{beta.mle}; \code{bfmd.pdf}
}
