% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmc.R
\name{bmc.prior.default}
\alias{bmc.prior.default}
\title{Default BMC Prior}
\usage{
bmc.prior.default(kth)
}
\arguments{
\item{kth}{the \out{<i>k</i><sup>th</sup>} seismic station used for distance calculation (if not provided,
\code{kth = 4})}
}
\value{
the BMC prior parameter list
\itemize{
\item \code{c1}, \code{c2}, \code{c3} the empirical parameters
\item \code{sigma} the standard error
\item \code{kth}   the kth seismic station used for distance calculation
\item \code{support}   the information supporting the prior model (here
\code{support = "default"})
}
}
\description{
List the parameters of the default prior model of the Bayesian Magnitude of
Completeness (BMC) method, as defined in Mignan et al. (2011) for different
\out{<i>k</i><sup>th</sup>} values (only \code{kth = 3}, \code{4} and \code{5}
allowed, otherwise returns \code{NULL}).
}
\details{
The generic model is defined as the prior model evaluated for the Taiwan earthquake data
(Mignan et al., 2011), as it represents the best constrained data set so far (read more
on this in Mignan and Chouliaras, 2014). It often provides a better prior once calibrated
to other data than a new fit to the data (see example given for the function
\code{bmc.prior}). It is also used for rapid mapping of the optimal
\emph{\out{m<sub>c</sub>}} map (see function \code{mc.geogr}).
}
\references{
Mignan, A., Werner, M.J., Wiemer, S., Chen, C.-C., Wu, Y.-M. (2011),
Bayesian Estimation of the Spatially Varying Completeness Magnitude of Earthquake
Catalogs, Bull. Seismol. Soc. Am., 101, 1371-1385,
\href{https://pubs.geoscienceworld.org/bssa/article-lookup/101/3/1371}{doi: 10.1785/0120100223}

Mignan, A., Chouliaras, G. (2014), Fifty Years of Seismic Network
Performance in Greece (1964-2013): Spatiotemporal Evolution of the Completeness Magnitude,
Seismol. Res. Lett., 85, 657-667
\href{https://pubs.geoscienceworld.org/ssa/srl/article-abstract/85/3/657/315375/fifty-years-of-seismic-network-performance-in}{doi: 10.1785/0220130209}
}
\seealso{
\code{bmc}; \code{bmc.prior}; \code{mc.geogr}
}
