% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmc.R
\name{bmc.prior.generic}
\alias{bmc.prior.generic}
\title{Generic BMC Prior}
\usage{
bmc.prior.generic(kth)
}
\arguments{
\item{kth}{the kth seismic station used for distance calculation (if not provided,
\code{kth = 4})}
}
\value{
the BMC prior parameter list
\itemize{
\item \code{c1}, \code{c2}, \code{c3} the empirical parameters
\item \code{sigma} the standard error
\item \code{kth}   the kth seismic station used for distance calculation
\item \code{support}   the information supporting the prior model
}
}
\description{
List the parameters of the generic prior model of the Bayesian Magnitude of
Completeness (BMC) method, as defined in Mignan et al. (2011) for different kth values
(only \code{kth = 3}, \code{4} and \code{5} allowed, otherwise returns \code{NULL}).
}
\details{
The generic model is defined as the prior model evaluated for the Taiwan earthquake data
(Mignan et al., 2011), as it represents the best constrained data set so far (read more
on this in refs). It oftens provides a better prior once calibrated to other data than a
new fit to the data (see example given for the function \code{bmc.prior}). It is also used
for rapid mapping of the optimal mc map (see function \code{mc.geogr}).
}
\references{
Mignan, A., Werner, M.J., Wiemer, S., Chen, C.-C., Wu, Y.-M. (2011),
Bayesian Estimation of the Spatially Varying Completeness Magnitude of Earthquake
Catalogs, Bull. Seismol. Soc. Am., 101, 1371-1385,
\href{https://pubs.geoscienceworld.org/bssa/article-lookup/101/3/1371}{doi: 10.1785/0120100223}
}
\seealso{
\code{bmc}; \code{bmc.prior}; \code{mc.geogr}
}
