% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmc.R
\name{bmc.prior}
\alias{bmc.prior}
\title{BMC Prior}
\usage{
bmc.prior(mc.obs, stations, kth = 4, support = "calibrated",
  dist.calc = "fast")
}
\arguments{
\item{mc.obs}{a data frame of the observed \out{m<sub>c</sub>} map defined by the
function \code{mc.geogr}}

\item{stations}{the seismic network data frame of parameters:
\itemize{
\item \code{lon} the seismic station longitude
\item \code{lat} the seismic station latitude
\item \code{...} other station attributes
}}

\item{kth}{the \out{k<sup>th</sup>} nearest seismic station used for distance
calculation (if not provided, \code{kth = 4})}

\item{support}{the information supporting the prior model: \code{"calibrated"} or
\code{"data"} (if not provided, \code{support = "calibrated"} - see Details)}

\item{dist.calc}{the method to be used to evaluate distances (if not provided,
\code{dist.calc = "fast"}; read Details of function \code{d.geogr2km})}
}
\value{
A list of:

the BMC prior model parameter list:
\itemize{
\item \code{c1}, \code{c2}, \code{c3} the empirical parameters
\item \code{sigma} the standard deviation
\item \code{kth}   the \out{k<sup>th</sup>} nearest seismic station used for distance calculation
\item \code{support}   the information supporting the prior model
}

the input data frame:
\itemize{
\item \code{mc}   the completeness magnitude value per cell
\item \code{d.kth}   the distance to the \out{k<sup>th</sup>} nearest seismic station per
cell (in km)
}
}
\description{
Define the prior model of the Bayesian Magnitude of Completeness (BMC) method (Mignan et
al., 2011) by fitting a function of the form
\ifelse{html}{\out{m<sub>c</sub> = c<sub>1</sub>d^c<sub>2</sub>+c<sub>3</sub>}}{\eqn{m_c = c_1 d^{c_2} +c_3}}
the observed completeness magnitude per cell and d the distance from the cell center to
the \out{k<sup>th</sup>} nearest seismic station.
}
\details{
\code{support = "calibrated"} uses the default BMC prior model defined by the function
\code{bmc.prior.default} and calibrates it to the \code{mc.obs} data by shifting the
residual average to 0, substracting it from \out{c<sub>3</sub>} (see e.g., Mignan et al., 2013; Mignan
and Chouliaras, 2014).

\code{support = "data"} directly fits the prior function to the \code{mc.obs} data
by using the Nonlinear Least Squares function \code{stats::nls}. If the estimation fails,
\code{support = "calibrated"} is used instead.
}
\examples{
# download the Southern California relocated catalogue of Hauksson et al. (2012)
url <- "http://service.scedc.caltech.edu/ftp/catalogs/"
cat <- "hauksson/Socal_DD/hs_1981_2011_06_comb_K2_A.cat_so_SCSN_v01"
dat <- scan(paste(url, cat, sep = ""), what = "character", sep = "\\n")
yr <- as.numeric(substr(dat, start=1, stop=4))
lat <- as.numeric(substr(dat, start=35, stop=42))
lon <- as.numeric(substr(dat, start=44, stop=53))
m <- as.numeric(substr(dat, start=63, stop=67))
seism <- data.frame(yr = yr, lon = lon,lat = lat, m = m)

# download the Southern California seismic network data
url <- "http://service.scedc.caltech.edu/station/weblist.php"
dat <- scan(url, what = "character", sep = "\\n", skip = 7)
network <- substr(dat, start = 1, stop = 2)
sta.name <- substr(dat, start = 5, stop = 9)
sta.lat <- as.numeric(substr(dat, start = 52, stop = 59))
sta.lon <- as.numeric(substr(dat, start = 61, stop = 70))
sta.on <- as.numeric(substr(dat, start = 78, stop = 81))
sta.off <- as.numeric(substr(dat, start = 89, stop = 92))
stations <- data.frame(lon = sta.lon, lat = sta.lat, name = sta.name)
stations <- subset(stations, (network == "CI" & sta.off > min(seism$yr) & sta.on < max(seism$yr)))
stations <- subset(stations, (duplicated(name) == F))

# map the observed mc (this may take a few minutes)
mc.obs <- mc.geogr(seism, "mode", "grid", dbin = 0.1)

# test the two possible priors & plot
model.calibrated <- bmc.prior(mc.obs, stations, kth = 5)
model.fromdata <- bmc.prior(mc.obs, stations, kth = 5, support = "data")
data <- model.calibrated[[2]]
di <- seq(0, max(data$d.kth))
params.cal <- model.calibrated[[1]]
params.dat <- model.fromdata[[1]]
plot(data$d.kth, data$mc.obs)
lines(di, params.cal$c1*di^params.cal$c2+params.cal$c3, col="orange")
lines(di, params.dat$c1*di^params.dat$c2+params.dat$c3, col="red")
}
\references{
Mignan, A., Werner, M.J., Wiemer, S., Chen, C.-C., Wu, Y.-M. (2011),
Bayesian Estimation of the Spatially Varying Completeness Magnitude of Earthquake
Catalogs, Bull. Seismol. Soc. Am., 101, 1371-1385,
\href{https://pubs.geoscienceworld.org/bssa/article-lookup/101/3/1371}{doi: 10.1785/0120100223}

Mignan, A., Jiang, C., Zechar, J.D., Wiemer, S., Wu, Z., Huang, Z. (2013),
Completeness of the Mainland China Earthquake Catalog and Implications for the Setup of
the China Earthquake Forecast Texting Center, Bull. Seismol. Soc. Am., 103, 845-859,
\href{https://pubs.geoscienceworld.org/ssa/bssa/article/103/2A/845/331723/completeness-of-the-mainland-china-earthquake}{doi: 10.1785/0120120052}

Mignan, A., Chouliaras, G. (2014), Fifty Years of Seismic Network
Performance in Greece (1964-2013): Spatiotemporal Evolution of the Completeness Magnitude,
Seismol. Res. Lett., 85, 657-667
\href{https://pubs.geoscienceworld.org/ssa/srl/article-abstract/85/3/657/315375/fifty-years-of-seismic-network-performance-in}{doi: 10.1785/0220130209}
}
\seealso{
\code{bmc}; \code{bmc.prior.default}; \code{d.geogr2km}; \code{mc.geogr}
}
