% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmc.R
\name{d.geogr}
\alias{d.geogr}
\title{Map the completeness magnitude in space}
\usage{
d.geogr(lon1, lat1, lon2, lat2)
}
\arguments{
\item{m}{Earthquake magnitude vector}
}
\description{
Map the completeness magnitude mc in space for a seismicity data frame, following
one of two possible geographical mappings: "grid" (computes mc in cells of size dbin)
or "bmc" (computes mc according to the Bayesian Magnitude of Completeness method of
Mignan et al., Bull. Seismol. Soc. Am., 2011).
}
