% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmc.R
\name{d.geogr2km}
\alias{d.geogr2km}
\title{Distance between Geographical Points}
\usage{
d.geogr2km(pt.ref, pt.list, method)
}
\arguments{
\item{pt.ref}{a list of 2 parameters:
\itemize{
\item \code{lon} the reference location longitude
\item \code{lat} the reference location latitude
}}

\item{pt.list}{a list or data frame of 2 parameters:
\itemize{
\item \code{lon} the other location longitude(s)
\item \code{lat} the other location latitude(s)
}}

\item{method}{the method to be used to evaluate d: \code{"fast"} or \code{"haversine"}
(see Details)}
}
\value{
the value(s) of the distance d between \code{pt.ref} and \code{pt.list}.
}
\description{
Calculate the distance d (in kilometers) between two geographical points or between one
geographical point and a vector of geographical points.
}
\details{
The \code{method = "fast"} assumes that the surface is flat. The
\code{method = "haversine"}, slower, calculates the haversine distance by using the
function \code{geosphere::distHaversine}. For regional earthquake catalogues, the error
of the fast method is insignificant.
}
