% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmd.R
\name{efmd.pdf}
\alias{efmd.pdf}
\title{Compute the pdf of the elemental FMD}
\usage{
efmd.pdf(m, theta)
}
\arguments{
\item{m}{a numeric vector of earthquake magnitudes}

\item{mc}{the ompleteness magnitude number}

\item{mbin}{Magnitude binning number}
}
\value{
The value of beta for a given m vector
}
\description{
Compute the pdf of the elemental FMD
}
\examples{
...
}
