% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmd.R
\name{efmd.sim}
\alias{efmd.sim}
\title{Elemental FMD Simulation}
\usage{
efmd.sim(N, theta, mbin = 0.1)
}
\arguments{
\item{N}{number of earthquakes}

\item{mbin}{Magnitude binning number}

\item{mc}{Completeness magnitude number}
}
\value{
A numeric vector of \code{N} earthquake magnitudes.
}
\description{
Simulate the elemental earthquake frequency magnitude distribution (eFMD) by
applying the Inverse Method to the angular FMD model (Mignan, 2012)
with probability density function \code{efmd.pdf}.
}
\details{
If \code{mbin} is not specified it assumes the default value of \code{0.1}.

The angular FMD model has density
}
\examples{
theta <- list(kappa = 3 * log(10), beta = log(10), mc = 2)
m.sim <- efmd.sim(1e3, theta)
mdistr <- fmd(m.sim)
plot(mdistr$mi, mdistr$Ni, log = "y")
points(mdistr$mi, mdistr$ni)
}
\references{
Mignan, A. (2012), Functional shape of the earthquake frequency-magnitude
distribution and completeness magnitude, J. Geophys. Res., 117, B08302,
\href{http://onlinelibrary.wiley.com/doi/10.1029/2012JB009347/full}{doi: 10.1029/2012JB009347}
}
\seealso{
\code{efmd.pdf} for the probability density function of the eFMD.
}
