% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_beta}
\alias{get_beta}
\title{Compute the FMD beta-value}
\usage{
get_beta(m, mc, mbin)
}
\arguments{
\item{m}{Earthquake magnitude vector}

\item{mc}{Completeness magnitude number}

\item{mbin}{Magnitude binning number}
}
\value{
The value of beta for a given m vector
}
\description{
Compute the FMD beta-value
}
\examples{
beta <- log(10); mc <- 2; mbin <- 0.1
m <- round(mc - mbin / 2 + rexp(100, beta), digits = log10(1/mbin))
get_beta(m, mc, mbin)
}
