% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmc.R
\name{mc.geogr}
\alias{mc.geogr}
\title{Completeness Magnitude Spatial Mapping}
\usage{
mc.geogr(seism, method, mapping, mbin = 0.1, box = NULL, dbin = NULL,
  nmin = 4)
}
\arguments{
\item{seism}{an earthquake catalogue data frame of parameters:
\itemize{
\item \code{lon} the earthquake longitude
\item \code{lat} the earthquake latitude
\item \code{m}   the earthquake magnitude
}}

\item{mbin}{the magnitude binning value (if not provided, \code{mbin = 0.1})}

\item{box}{a vector of the minimum longitude, maximum longitude, minimum latitude and
maximum latitude, in this order}

\item{dbin}{the spatial binning value (if not provided, \code{dbin} is calculated such that
the map is made of 5 longitudinal cells based on \code{box})}
}
\value{
The completeness magnitude mc data frame
}
\description{
Map the completeness magnitude mc in space for a seismicity data frame, following
one of two possible geographical mappings: "grid" (computes mc in cells of size dbin)
or "bmc" (computes mc according to the Bayesian Magnitude of Completeness method of
Mignan et al., Bull. Seismol. Soc. Am., 2011).
}
\examples{
mbin <- 0.1
url <- "http://service.scedc.caltech.edu/ftp/catalogs/"
dat <- "hauksson/Socal_DD/hs_1981_2016_comb_K4_A.cat_so_SCSN_v2q"
seism <- scan(paste(url, dat, sep = ""), what = "character", sep = "\\n")
yr <- as.numeric(substr(seism, start=1, stop=4))
lat <- as.numeric(substr(seism, start=35, stop=42))
lon <- as.numeric(substr(seism, start=44, stop=53))
mag <- round(as.numeric(substr(seism, start=63, stop=67)), digits = log10(1/mbin))
seism <- data.frame(yr = yr, lon = lon,lat = lat, m = mag)
seism <- subset(seism, yr >= 2010)
mc.obs <- mc.geogr(seism, "mode", "grid", dbin = 0.1)
image(matrix(mc.obs$mc.obs, nrow=length(unique(mc.obs$lon)), ncol=length(unique(mc.obs$lat))))
}
