% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmd.R
\name{beta.mle}
\alias{beta.mle}
\title{Compute the FMD beta-value}
\usage{
beta.mle(m, mc, mbin)
}
\arguments{
\item{m}{a numeric vector of earthquake magnitudes}

\item{mc}{Completeness magnitude number}

\item{mbin}{Magnitude binning number}
}
\value{
The value of beta for a given m vector
}
\description{
Compute the FMD beta-value
}
\examples{
beta <- log(10); mc <- 2; mbin <- 0.1
m <- round(mc - mbin / 2 + rexp(100, beta), digits = log10(1/mbin))
get_beta(m, mc, mbin)
}
