% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmc.R
\name{bmc.prior.default}
\alias{bmc.prior.default}
\title{Default BMC Prior}
\usage{
bmc.prior.default(kth)
}
\arguments{
\item{kth}{the \out{k<sup>th</sup>} nearest seismic station used for
distance calculation (if not provided, \code{kth = 4})}
}
\value{
the BMC prior parameter list
\itemize{
\item \code{c1}, \code{c2}, \code{c3} the empirical parameters
\item \code{sigma} the standard error
\item \code{kth}   the \out{k<sup>th</sup>} nearest seismic station used for
distance calculation
\item \code{support}   the information supporting the prior model (here
\code{support = "default"})
}
}
\description{
List the parameters of the default prior model of the Bayesian Magnitude of
Completeness (BMC) method, as defined in Mignan et al. (2011) for different
\out{k<sup>th</sup>} values (only \code{kth = 3}, \code{4} and \code{5}
allowed, otherwise returns \code{NULL}).
}
\details{
The BMC default model is defined as the prior model evaluated for the Taiwan earthquake data
(Mignan et al., 2011), as it represents the best constrained data set so far (read more
on this in Mignan and Chouliaras, 2014). It often provides a better prior once calibrated
to other data than a new fit to the data (see example given for the function
\code{bmc.prior}). It is also used for rapid mapping of the optimal
\out{m<sub>c</sub>} map (see function \code{mc.geogr}).
}
\examples{
# map the predicted mc for a set of simulated stations
box <- c(-5, 5, -5, 5); dbin <- 0.1  #degrees
sta.n <- 30
stations <- data.frame(lon = rnorm(sta.n), lat=rnorm(sta.n))
grid <- expand.grid(lon = seq(box[1], box[2], dbin), lat = seq(box[3], box[4], dbin))
grid.n <- nrow(grid)

kth <- 4
params <- bmc.prior.default(kth)
d <- sapply(1:grid.n, function(i) rseismNet::d.geogr2km(grid[i,], stations, method = "fast"))
d.kth <- sapply(1:grid.n, function(i) sort(d[,i])[kth])
mc.pred <- (params$c1 * d.kth ^ params$c2 + params$c3)
image(unique(grid$lon), unique(grid$lat),
  matrix(mc.pred, nrow=length(unique(grid$lon)), ncol=length(unique(grid$lat))))
points(stations, pch = 2)
}
\references{
Mignan, A., Werner, M.J., Wiemer, S., Chen, C.-C., Wu, Y.-M. (2011),
Bayesian Estimation of the Spatially Varying Completeness Magnitude of Earthquake
Catalogs, Bull. Seismol. Soc. Am., 101, 1371-1385,
\href{https://pubs.geoscienceworld.org/bssa/article-lookup/101/3/1371}{doi: 10.1785/0120100223}

Mignan, A., Chouliaras, G. (2014), Fifty Years of Seismic Network
Performance in Greece (1964-2013): Spatiotemporal Evolution of the Completeness Magnitude,
Seismol. Res. Lett., 85, 657-667
\href{https://pubs.geoscienceworld.org/ssa/srl/article-abstract/85/3/657/315375/fifty-years-of-seismic-network-performance-in}{doi: 10.1785/0220130209}
}
\seealso{
\code{bmc}; \code{bmc.prior}; \code{mc.geogr}
}
