% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmd.R
\name{efmd.sim}
\alias{efmd.sim}
\title{Simulation of the Elemental FMD}
\usage{
efmd.sim(N, theta, mbin = 0.1)
}
\arguments{
\item{N}{the number of earthquakes to simulate}

\item{theta}{a list of 3 parameters:
\itemize{
\item \code{kappa} the detection parameter value
\item \code{beta}  the Gutenberg-Richter parameter value
\item \code{mc}    the completeness magnitude value
}}

\item{mbin}{the magnitude binning value (if not provided, \code{mbin = 0.1})}
}
\value{
A numeric vector of \code{N} earthquake magnitudes.
}
\description{
Simulate the elemental earthquake frequency magnitude distribution (eFMD) by
applying the Inversion Method (Devroye, 1986) to the angular FMD model of
Mignan (2012).
}
\details{
The angular FMD model is defined as an Asymmetric Laplace distribution. It has an
angular shape in the log-lin space and corresponds to the case where the completeness
magnitude \out{m<sub>c</sub>} is constant (read more in Mignan, 2012; Mignan and
Chen, 2016).
}
\examples{
theta <- list(kappa = 3 * log(10), beta = log(10), mc = 2)
m.sim <- efmd.sim(1e4, theta)
mdistr <- fmd(m.sim)
plot(mdistr$mi, mdistr$Ni, log = "y")
points(mdistr$mi, mdistr$ni)
}
\references{
Devroye, L. (1986), Non-Uniform Random Variate Generation,
Springer-Verlag New York Inc., New York, 843 pp.

Mignan, A. (2012), Functional shape of the earthquake frequency-magnitude
distribution and completeness magnitude, J. Geophys. Res., 117, B08302,
\href{http://onlinelibrary.wiley.com/doi/10.1029/2012JB009347/full}{doi: 10.1029/2012JB009347}

Mignan, A., Chen, C.-C. (2016), The Spatial Scale of Detected Seismicity,
Pure Appl. Geophys., 173, 117-124,
\href{https://link.springer.com/article/10.1007/s00024-015-1133-7}{doi: 10.1007/s00024-015-1133-7}
}
\seealso{
\code{beta.mle}; \code{mc.val}
}
