% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmd.R
\name{fmd}
\alias{fmd}
\title{Compute the earthquake FMD}
\usage{
fmd(m, mbin)
}
\arguments{
\item{m}{Earthquake magnitude vector}

\item{mbin}{Magnitude binning number}
}
\value{
The earthquake frequency magnitude distribution (FMD)
}
\description{
Compute the earthquake FMD
}
\examples{
beta <- log(10); mc <- 2; mbin <- 0.1
m <- mc - mbin / 2 + rexp(100, beta)
mdistr <- fmd(m, mbin)
plot(mdistr$mi, mdistr$Ni, log = "y")
points(mdistr$mi, mdistr$ni)
}
