% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmc.R
\name{mc.geogr}
\alias{mc.geogr}
\title{Map the completeness magnitude in space}
\usage{
mc.geogr(seism, mbin, method, mapping, dbin = NULL, box = NULL, nmin = 4)
}
\arguments{
\item{mbin}{Magnitude binning number}

\item{m}{Earthquake magnitude vector}

\item{lon}{Earthquake longitude vector}

\item{lat}{Earthquake latitude vector}
}
\value{
The completeness magnitude mc data frame
}
\description{
Map the completeness magnitude mc in space for a seismicity data frame, following
one of two possible geographical mappings: "grid" (computes mc in cells of size dbin)
 or "bmc" (computes mc according to the Bayesian Magnitude of Completeness method of
 Mignan et al., Bull. Seismol. Soc. Am., 2011).
}
\examples{
mbin <- 0.1
url <- "http://service.scedc.caltech.edu/ftp/catalogs/"
dat <- "hauksson/Socal_DD/hs_1981_2016_comb_K4_A.cat_so_SCSN_v2q"
seism <- scan(paste(url, dat, sep = ""), what = "character", sep = "\\n")
yr <- as.numeric(substr(seism, start=1, stop=4))
lat <- as.numeric(substr(seism, start=35, stop=42))
lon <- as.numeric(substr(seism, start=44, stop=53))
mag <- round(as.numeric(substr(seism, start=63, stop=67)), digits = log10(1/mbin))
seism <- data.frame(yr = yr, lon = lon,lat = lat, mag = mag)
seism <- subset(seism, yr >= 2010)
mc <- mc.geogr(seism, mbin, "mode", "grid", 0.1)
image(matrix(mc$mc, nrow=length(unique(mc$lon)), ncol=length(unique(mc$lat))))
}
