% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmd.R
\name{mc.val}
\alias{mc.val}
\title{Estimate the completeness magnitude}
\usage{
mc.val(m, mbin, method)
}
\arguments{
\item{m}{Earthquake magnitude vector}

\item{mbin}{Magnitude binning number}

\item{method}{Method following which the completeness magnitude mc is computed.
Options include "mode" and "KStest"}
}
\value{
The completeness magnitude mc
}
\description{
Estimate the completeness magnitude mc from the distribution of the magnitude m
using two different methods, the mode or a KS test of the Gutenberg-Richter law.
}
\examples{
m <- 
mc <- mc.val(m, 0.1, "mode")
}
